# nextjs-frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ingest/documents": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest Documents",
        "operationId": "ingest_documents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestDocument": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "\u6587\u6863ID"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "\u6587\u6863\u6765\u6e90"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "\u6587\u6863\u6807\u9898"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "\u6587\u6863\u6b63\u6587\u6587\u672c"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source",
          "title",
          "content"
        ],
        "title": "IngestDocument"
      },
      "IngestRequest": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/IngestDocument"
            },
            "type": "array",
            "title": "Documents"
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "IngestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}